+++ CHANGELOG.md
@@ -3,6 +3,18 @@ All notable changes to this project will be documented in this file. This change
 Discljord follows semantic versioning.
 
 ## [Unreleased]
+### Added
+ - Support for audit log reasons on all requests made with `discljord.messaging`
+ - Support for gateway intents
+ - Support for allowed-mentions in `discljord.messaging/execute-webhook!` payload
+ - Support for allowed-mentions in `discljord.messaging/create-message!` payload
+
+### Changed
+ - `discljord.messaging/execute-webhook!` had its parameters moved to optional ones
+
+### Fixed
+ - `discljord.messaging/execute-webhook!` was unable to send files
+ - Documentation for create-message! saying file sharing was not implemented
 
 ## [0.2.7]
 ### Fixed